change:chart.jsの切り分け
