+++ 作成にあたって/本リリースのための参考にした文献
@@ -21,3 +21,64 @@ https://zenn.dev/yukihaga/articles/3d49208638e397
 kaminari
 
 daisyUIのimport 少し問題あり。改善方法今の所不明。警告出ていているが特に問題ないので一旦放置。
+
+
+目的（再確認）
+
+チーム全体で共有するカレンダーに、誰がどの日にトレーニングしたかを表示する
+
+表示対象：今月など約1ヶ月分のカレンダー
+1日に複数人いれば〇を複数並べる
+表示は「苗字だけ」
+ポジションでフィルター可能
+特定選手の履歴表示時は名前を画面に明示
+📅 全体の流れ（7ステップ）
+
+【ステップ 1】FullCalendarをRailsに導入
+
+Yarn経由で @fullcalendar/core, @fullcalendar/daygrid などをインストール
+カレンダーを表示する画面用のJSファイル（またはStimulus controller）を作成
+カレンダーを埋め込むためのHTML <div> を用意
+【ステップ 2】Rails側APIを用意する
+
+/api/calendar/trainings のようなエンドポイントを作成（Api::CalendarControllerなど）
+現在ログイン中のチームに所属する全ユーザーの、トレーニング履歴を取得
+日付・苗字・ユーザーID（選手をクリックするため）などを含むJSONを返す
+【ステップ 3】カレンダー画面にイベント（〇＋苗字）を描画
+
+FullCalendarの eventData にAPIで取得したJSONを流し込む
+eventRender（または eventContent）をカスタムして、苗字を〇で囲んで表示
+日付ごとに複数ユーザーがいれば、複数のイベントをその日に並べて表示
+【ステップ 4】ポジションフィルター機能を追加
+
+上部に <select> タグでポジションを選べるUIを追加
+選択されたポジションに該当するユーザーだけを表示するようイベントを再描画
+選択が「すべて」の場合は全ユーザーを対象に表示
+【ステップ 5】個人カレンダーモードを用意（オプション）
+
+URLに ?user_id=◯ のようなパラメータがついているときは、その選手だけの表示に切り替える
+ページ上部に「◯◯さんのトレーニング履歴」とテキスト表示
+それ以外の時はチーム全体のカレンダーとして振る舞う
+【ステップ 6】デザイン調整（〇の中に苗字表示）
+
+Tailwind CSS + カスタムCSSで、
+丸く表示（rounded-full）
+苗字を中央に
+同じ日に複数丸が並んでも崩れないように調整
+【ステップ 7】将来的な拡張も考慮
+
+イベントをクリックしてトレーニング詳細ページへ遷移
+月切り替え時の再フェッチ（必要なら）
+曜日表示、日本語化など
+🧩 構成イメージ
+
+Model: Training, User, Position, Team
+Controller: Api::CalendarController
+View: カレンダー用の専用ページ（例: calendar/index）
+JS/Stimulus: calendar_controller.js で FullCalendar を制御
+CSS: Tailwind で〇表示調整
+✅ これが終わると…
+
+「4月30日に久保・佐藤がトレーニング」→ カレンダー上に〇が2つ並ぶ
+ポジションで「WR」を選べばWRだけの丸が表示される
+久保選手の個人ページから「履歴を見る」を押すと、その人だけのカレンダー表示
